requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
awswrangler==3.5.0
boto3==1.34.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pyyaml==6.0.1
faker==20.1.0
python-dateutil==2.8.2
requests==2.31.0
aiofiles==23.2.1 